use a component with lifecycle methods to fetch todos
We're using `ReactRouter.withRouter` to wrap the component generated by
`ReactRedux.connect`, so on route change, the wrapped component will
receive the new route via props, and the `componentDidUpdate` lifecycle
hook will run.

However, we can't have hooks for function components, so we define a
simple class. In it's `render`, we just return `TodoList`, passing it the
props. We re-assign to the same variable this class, wrapped in
`withRouter` and then `connect`.