add promise support for dispatch
Instead of waiting for the promise in component lifecycle methods, we
monkey-patch `store.dispatch` so that it checks whether the `action` it
receives as a parameter is a promise. If it is, it will wait for it to
resolve, and call the original `dispatch` with the result. Otherwise it
just calls the original `dispatch` with `action` as a param.

So the action which would upade the state with async data, such as AJAX
responses, will get dispatched once the async promise is resolved.
Because components are subscribed to the store via `connect`, they will
re-render, using the updated state.