display loading indicator
Helped groking top-level and specific selectors. In the same file as
the top-level reducer, we define top-level selectors - which call the
specific selectors, giving them the appropriate slice of state - and the
specific selectors assume they're getting their slice, instead of the
global state object. So we can freely modify the sub-reducers - as long
as we update the respective sub-selectors, the app will continue to work
and require no changes in the main reducer or in the components.