use redux-thunk middleware
Similar to express middleware, Redux actions will be passed through each
middleware function; these middlewares can do anything they want, but the
contract is they will call 'next' with an action object.

The thunk middleware is simple; it just checks whether the action is a
function - if yes, it calls that function, passing it the `dispatch` and
`getState` functions.

`createStore` takes, besides the root reducer, an 'enhancer' function.
So if an enhancer is passed, then `createStore` will first call it, giving
it a reference to itself, and call the result (enhancers return functions)
with the root reducer and (optional) preloaded state.