pass err handler as second param to `then`
If a promise `catch` function is used, then it will catch errors from the
success handler, if any:
Example:

  api.doSomething().then(res => {
    throw new Error('foobar')
    sendResponse(res)
  }).catch(error => {
    sendError(error.message)
  })

User sees 'foobar'

These are unexpected 'programmer errors' and therefore we don't know what
kind of error message we would get - so we don't want to show it to the
user:

  api.doSomething().then(res => {
    throw new Error('foobar')
    sendResponse(res)
  }, error => {
    sendError(error.message)
  })

In this case, the exception won't be handled by the error handler passed
as the second param to `then`.  However, this effectively swallows the
exception, leaving the app in an undefined state and sending no response
at all.

The examples are obviously simplified; in a real app, I think the best
course of action is to send back a generic 500, and log the exception with
a service like Sentry. So, we protect users from internal errors, while
making them available for analysis:

  api.doSomething().then(res => {
    try {
      throw new Error('foobar')
      sendResponse(res)
    } catch (internalException) {
      sendSentryReport(internalException)
      sendError('something broke.') // generic error message
    }
  }, error => {
    sendError(error.message)
  })